+++ shooney/src/main/java/com/shun/blog/service/portfolio/music/MusicService.java
@@ -6,13 +6,13 @@ import com.shun.blog.model.portfolio.music.Music;
 
 public interface MusicService {
 	
-	Music findById(int id);
+	Music findById(Long id);
 	
 	void saveMusic(Music music);
 	
 	void updateMusic(Music music);
 	
-	void deleteUserById(int id);
+	void deleteUserById(Long id);
 
 	List<Music> findAllMusics(); 
 }
